Frontend: Chatbot entegrasyonu tamamlandi! AI servisi (Python) ile arayuz (React) basariyla birlestirildi. Tum sistemler aktif.
