--- Chatbot.css
@@ -1,69 +1,0 @@
-.chatbot-container {
-    position: fixed;
-    bottom: 90px;
-    right: 30px;
-    width: 350px;
-    height: 450px;
-    background: #ffffff;
-    box-shadow: 0px 0px 15px rgba(0, 0, 0, 0.2);
-    border-radius: 12px;
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-    z-index: 9999;
-}
-
-.chatbot-header {
-    background: #2563eb;
-    color: white;
-    padding: 12px;
-    display: flex;
-    justify-content: space-between;
-    font-size: 1.1rem;
-}
-
-.chatbot-messages {
-    flex: 1;
-    padding: 10px;
-    overflow-y: auto;
-}
-
-.msg {
-    padding: 8px 10px;
-    margin-bottom: 8px;
-    border-radius: 8px;
-    max-width: 80%;
-}
-
-.msg.user {
-    background: #dbeafe;
-    margin-left: auto;
-}
-
-.msg.bot {
-    background: #f3f4f6;
-    margin-right: auto;
-}
-
-.chatbot-input {
-    display: flex;
-    padding: 10px;
-    border-top: 1px solid #ddd;
-}
-
-.chatbot-input input {
-    flex: 1;
-    padding: 8px;
-    border: 1px solid #ddd;
-    border-radius: 6px;
-    margin-right: 8px;
-}
-
-.chatbot-input button {
-    background: #2563eb;
-    color: white;
-    border: none;
-    padding: 8px 12px;
-    border-radius: 6px;
-    cursor: pointer;
-}
