+++ ai-module/requirements.txt
@@ -1,0 +1,4 @@
+fastapi
+uvicorn
+requests
+python-dotenv
